Renamed vars and method names as per PR review suggestions
