+++ frontend/server/src/main/java/com/amazonaws/ml/mms/ModelServer.java
@@ -15,7 +15,7 @@ package com.amazonaws.ml.mms;
 import com.amazonaws.ml.mms.archive.ModelArchive;
 import com.amazonaws.ml.mms.archive.ModelException;
 import com.amazonaws.ml.mms.metrics.MetricManager;
-import com.amazonaws.ml.mms.servingsdk_impl.PluginsManager;
+import com.amazonaws.ml.mms.servingsdk.impl.PluginsManager;
 import com.amazonaws.ml.mms.util.ConfigManager;
 import com.amazonaws.ml.mms.util.Connector;
 import com.amazonaws.ml.mms.util.ServerGroups;
