+++ serving-sdk/src/main/java/software/amazon/ai/mms/servingsdk/Worker.java
@@ -16,11 +16,11 @@ public interface Worker {
      * Get the current running status of this model's worker
      * @return True - if the worker is currently running. False - the worker is currently not running.
      */
-    boolean getIsRunning();
+    boolean isRunning();
 
     /**
      * Get the current memory foot print of this worker
      * @return Current memory usage of this worker
      */
-    Long getWorkerMemory();
+    long getWorkerMemory();
 }
